@@ -1,4 +1,4 @@
-const { matcherTest, buildReturnMessage } = require('../utils')
+const { matcherTest, buildReturnMessage } = require('../utils');
 
 function toHaveStyleRule({ props: {style} }, property, expected) {
   const styles = Array.isArray(style) ? style.filter(x => x) : style
@@ -6,9 +6,7 @@ function toHaveStyleRule({ props: {style} }, property, expected) {
   /**
    * Convert style name to camel case (so we can compare)
    */
-  const camelCasedProperty = property.replace(/-(\w)/, (_, match) =>
-    match.toUpperCase()
-  )
+  const camelCasedProperty = property.replace(/-(\w)/, (_, match) => match.toUpperCase());
 
   /**
    * Merge all styles into one final style object and search for the desired
@@ -18,16 +16,13 @@ function toHaveStyleRule({ props: {style} }, property, expected) {
     Array.isArray(styles)
       ? styles.reduce((acc, item) => (Object.assign({}, acc, item)), {})
       : styles
-  const received = mergedStyles[camelCasedProperty]
-  const pass =
-    !received && !expected && this.isNot
-      ? false
-      : matcherTest(received, expected)
+  const received = mergedStyles[camelCasedProperty];
+  const pass = matcherTest(received, expected, this.isNot);
 
   return {
     pass,
     message: buildReturnMessage(this.utils, pass, property, received, expected),
-  }
+  };
 }
 
-module.exports = toHaveStyleRule
+module.exports = toHaveStyleRule;
